+++ app/components/primer/button_group.rb
@@ -7,10 +7,10 @@ module Primer
 
     # Required list of buttons to be rendered.
     #
-    # @param kwargs [Hash] The same arguments as <%= link_to_component(Primer::ButtonComponent) %> except for `variant` and `group_item`.
+    # @param kwargs [Hash] The same arguments as <%= link_to_component(Primer::ButtonComponent) %> except for `size` and `group_item`.
     renders_many :buttons, lambda { |**kwargs|
       kwargs[:group_item] = true
-      kwargs[:variant] = @variant
+      kwargs[:size] = @size
 
       Primer::ButtonComponent.new(**kwargs)
     }
@@ -25,19 +25,20 @@ module Primer
     #     <% component.button(classes: "custom-class") { "Custom class" } %>
     #   <% end %>
     #
-    # @example Variants
+    # @example Sizes
     #
-    #   <%= render(Primer::ButtonGroup.new(variant: :small)) do |component| %>
+    #   <%= render(Primer::ButtonGroup.new(size: :small)) do |component| %>
     #     <% component.button { "Default" } %>
     #     <% component.button(scheme: :primary) { "Primary" } %>
     #     <% component.button(scheme: :danger) { "Danger" } %>
     #     <% component.button(scheme: :outline) { "Outline" } %>
     #   <% end %>
     #
-    # @param variant [Symbol] <%= one_of(Primer::ButtonComponent::VARIANT_OPTIONS) %>
+    # @param variant [Symbol] DEPRECATED. <%= one_of(Primer::ButtonComponent::SIZE_OPTIONS) %>
+    # @param size [Symbol] <%= one_of(Primer::ButtonComponent::SIZE_OPTIONS) %>
     # @param system_arguments [Hash] <%= link_to_system_arguments_docs %>
-    def initialize(variant: Primer::ButtonComponent::DEFAULT_VARIANT, **system_arguments)
-      @variant = variant
+    def initialize(variant: nil, size: Primer::ButtonComponent::DEFAULT_SIZE, **system_arguments)
+      @size = variant || size
       @system_arguments = deny_tag_argument(**system_arguments)
       @system_arguments[:tag] = :div